adding aclSources for relays initial UI preview, api support, database migration

+++ pages/api/relay/[id]/blocklistpubkeys.ts
@@ -73,7 +73,7 @@ export default async function handle(req: any, res: any) {
             return;
         }
         // delete all lists with this reason..
-        if (listId != "") {
+        if (listId != "" && listId != "all") {
             await prisma.listEntryPubkey.deleteMany({
                 where: {
                     AND: [
